fix: wrap all floor() calls with abstractFloat args in f32() to prevent WGSL i32→f32 implicit conversion warnings
14 parametric .tsx files + 4 simple .ts files fixed: floor() with abstractFloat
arguments (numeric literals like 0.5, 2.0 mixed with f32) returns i32 on some
GPU drivers, causing implicit conversion warnings when used in f32 context.

Also added missing f32 import to apocarpetVar.ts and collatzVar.ts.

+++ packages/app/src/flame/variations/parametric/blur/blurPixelizeVar.tsx
@@ -42,8 +42,8 @@ export const blurPixelizeVar = parametricVariation(
   (pos, varInfo, P) => {
     'use gpu'
     const inv_size = 1.0 / P.size
-    const x0 = floor(pos.x * inv_size)
-    const y0 = floor(pos.y * inv_size)
+    const x0 = f32(floor(pos.x * inv_size))
+    const y0 = f32(floor(pos.y * inv_size))
     const cx = x0 * P.size + P.size * 0.5
     const cy = y0 * P.size + P.size * 0.5
     const dx = pos.x - cx
